// backend/nest-api/src/nest/modules/igt-serial-numbers/dto/update-igt-serial.dto.ts
export class UpdateIgtSerialDto {
  serial_number?: string;
  category?: string;
  status?: string;
  manufacturer?: string;
  model?: string;
  notes?: string;
  is_active?: number;
  updated_by?: string;
  used_at?: string;
  used_by?: string;
  used_in_asset_id?: number;
  used_in_asset_number?: string;
}